server/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
playwright==1.45.0
pluggy==1.4.0
psutil==6.0.0
psycopg==3.1.18
py==1.11.0
pyee==11.1.0
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
pytest-xprocess==0.23.0
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
typing_extensions==4.10.0
urllib3==2.2.2
Werkzeug==3.0.3

flask-sqlalchemy
flask-bcrypt
python-dotenv
psycopg2
flask-session